% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/araar.R
\name{araar_ci}
\alias{araar_ci}
\title{Bandas bootstrap para diferencia GLC (Araar)}
\usage{
araar_ci(
  yA,
  wA = NULL,
  yB,
  wB = NULL,
  p_grid = seq(0.05, 0.95, by = 0.01),
  B = 1000,
  conf = 0.95,
  seed = 123,
  progress = TRUE
)
}
\description{
Bandas bootstrap para diferencia GLC (Araar)
}
