% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/araar.R
\name{araar_test}
\alias{araar_test}
\title{Test de Araar (Dominancia SD2 por GLC)}
\usage{
araar_test(
  yA,
  wA = NULL,
  yB,
  wB = NULL,
  p_grid = seq(0.05, 0.95, by = 0.01),
  B = 1000,
  conf = 0.95
)
}
\description{
Test de Araar (Dominancia SD2 por GLC)
}
