% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glc.R
\name{glc_weighted}
\alias{glc_weighted}
\title{GLC ponderada (Generalized Lorenz Curve)}
\usage{
glc_weighted(y, w = NULL, p_grid = seq(0.01, 1, by = 0.01))
}
\description{
GLC ponderada (Generalized Lorenz Curve)
}
\keyword{internal}
