% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/araar.R
\name{plot_araar}
\alias{plot_araar}
\title{Gr치fico Araar din치mico (A vs B con colores)}
\usage{
plot_araar(araar_result, group_levels, title_prefix = "Test de Araar")
}
\description{
Gr치fico Araar din치mico (A vs B con colores)
}
