% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf_metric.R
\name{plot_cdf_metric}
\alias{plot_cdf_metric}
\title{plot_cdf_metric
genera gráfico CDF comparando grupo A vs grupo B}
\usage{
plot_cdf_metric(cdf_df, metric_name)
}
\description{
plot_cdf_metric
genera gráfico CDF comparando grupo A vs grupo B
}
